% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-other.R
\docType{data}
\name{bci_elevation}
\alias{bci_elevation}
\title{Elevation data from Barro Colorado Island (bci).}
\format{An object of class \code{list} of length 4.}
\usage{
bci_elevation
}
\description{
This dataset can be downloaded directly from
\url{http://bit.ly/fgeo-data-bci-elevation}.
}
\examples{
str(bci_elevation)
sapply(bci_elevation, class)
str(bci_elevation$col)
}
\keyword{datasets}
