% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-view.R
\docType{data}
\name{bci_taxa}
\alias{bci_taxa}
\title{Complete ViewTaxonomy (taxa) table from Barro Colorado Island (bci).}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1428 rows and 21 columns.}
\usage{
bci_taxa
}
\description{
This is a slightly cleaned version of a ViewTaxonomy table. For a raw .csv
version see \url{http://bit.ly/fgeo-opendata-taxa-bci}.
}
\examples{
str(bci_taxa, give.attr = FALSE)
}
\keyword{datasets}
