% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-view.R
\docType{data}
\name{bci_vft}
\alias{bci_vft}
\alias{bci_vft_1ha}
\alias{bci_vft_random}
\title{ViewFullTable (vft) from Barro Colorado Island (bci).}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10543 rows and 32 columns.}
\usage{
bci_vft_1ha

bci_vft_random
}
\description{
Slightly cleaned versions from a ViewFullTable table, from census 2005 and
census 2010 (see columns definition at \link{vft_description}):
\itemize{
\item \code{bci_vft_1ha}: All threes in one hectare (coordinates: \code{px >= 700}, \code{px < 800}, \code{py >=100}, \code{py < 200}). Raw .csv file available at
\url{http://bit.ly/fgeo-data-bci-vft-1ha}.
\item \code{bci_vft_random}: One thousand randomly-chosen trees (from the whole plot).
Raw .csv file available at \url{http://bit.ly/fgeo-data-bci-vft-random}).
}
}
\examples{
str(bci_vft_1ha, give.attr = FALSE)

str(bci_vft_random, give.attr = FALSE)

show_some_values_of_each_column <- function(x) head(unique(x))
lapply(bci_vft_random, show_some_values_of_each_column)
}
\seealso{
\link{vft_description}, \link{unique_id}.
}
\keyword{datasets}
