% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-census.R
\docType{data}
\name{census_bci}
\alias{census_bci}
\alias{stem6_1ha_bci}
\alias{stem7_1ha_bci}
\alias{tree6_1ha_bci}
\alias{tree7_1ha_bci}
\alias{stem6_random_bci}
\alias{stem7_random_bci}
\alias{tree6_random_bci}
\alias{tree7_random_bci}
\title{Tree and stem tables from Barro Colorado Island (bci).}
\format{See examples.}
\usage{
stem6_1ha_bci

stem7_1ha_bci

tree6_1ha_bci

tree7_1ha_bci

stem6_random_bci

stem7_random_bci

tree6_random_bci

tree7_random_bci
}
\description{
For a general description see \link{census_description}. These datasets are tree
and stem tables of the same trees selected to create the tables \link{vft_1ha_bci}
and \link{vft_random_bci}. In the name of the tree and stem tables documented
here, the number after \code{stem} or \code{census} matches a value of the column
\code{PlotCensusNumber} in the corresponding ViewFullTable.
}
\examples{
str(stem6_1ha_bci, give.attr = FALSE)

str(tree7_1ha_bci, give.attr = FALSE)

str(stem7_random_bci, give.attr = FALSE)

str(tree6_random_bci, give.attr = FALSE)

}
\seealso{
\link{vft_1ha_bci}, \link{vft_random_bci}, \link{census_description}.
}
\keyword{datasets}
