% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tree-stem.R
\name{census_description}
\alias{census_description}
\title{General description of census tables, tree and stem.}
\source{
\url{http://ctfs.si.edu/Public/CTFSRPackage/index.php/web/data_format}
}
\description{
Each tree and stem table -- collectively called \emph{census} tables -- is a
dataframes that stores data of one census of one plot. That is, for any given
plot one tree table stores data recorded during one specific census, and each
row records the data from one specific tree. And for any given plot one stem
table stores data recorded during one specific census , and each row records
the data from one specific stem. For example, if plot X was censused twice,
then it has two tree tables and two stem tables.
}
\details{
Each census table has two remarkable properties:
\itemize{
\item The number of rows equals the number of trees or stems that were ever found
in the plot (in any census). This implies that the number of rows is
identical to that of any other census table of the same plot; also, that
trees or stems appear in tree or stem tables before they recruit, while they
are alive, and after they die.
\item The order of the rows is identical to that of any other census table of the
same plot. This implies that you can compare data accross censuses even if
the row order is important for your analyses. Also, you can easily repeat
analyses with different censuses with minimal effort. For example, this
structure makes helps to calculate demographic rates between any pair of
censuses.
}
}
\section{Definition of Variables in tree tables}{

\itemize{
\item \code{treeID}: The unique tree identifier in CTFS database. Useful to be certain
in matching trees.
\item \code{stemID}: The unique stem identifier in CTFS database. Useful to be certain
in matching stems. FIXME: Missing in \code{tree6_1ha_bci}
\item \code{tag}: Tag number used in the field.
\item \code{StemTag}: Tag number on the individual stem, if present.
\item \code{sp}: The species mnemonic. To get full species names, the taxonomy table
must be downloaded from the CTFS database:
http://ctfs.arnarb.harvard.edu/CTFSReports
\item \code{quadrat}: Quadrat designation.
\item \code{gx}: The x coordinate within the plot, relative to one edge of the plot.
\item \code{gy}: The y plot coordinate.
\item \code{MeasureID}: The unique identifier of a single measurement in the CTFS
database. FIXME: Missing in \code{tree6_1ha_bci}
\item \code{CensusID}: The numeric identifier of the census.
\item \code{dbh}: Diameter of one stem on the tree, the stem whose stemID is given.
\item \code{pom}: The point-of-measure, where the diameter was taken, identical to hom,
but a character variable with only 2 decimal places.
\item \code{hom}: The height-of-measure, identical to pom but a numeric variable with
full precision.
\item \code{ExactDate}: The date on which a tree was measured.
\item \code{DFstatus}: The codes for the measurement as recorded in the field.
\itemize{
\item \code{A}: Alive.
\item \code{D}: Dead.
\item \code{lost_stem} (deprecated): Lost stem. It usually means the stem was
broken in the given census, while the tree had no other stem.  This
status is deprecated;  it is safer to check \code{stemID} to determine whether
a tree's measurement changed stems between censuses.
\item \code{M}: Missing. It is used where there is no record of the corresponding
value of \code{dbh}, so it is not certain whether the tree was alive or dead.
\code{P}: Prior. It indicates a tree had not yet recruited at this census.
}
\item \code{codes}: The codes for the measurement as recorded in the field.
\item \code{nostems}: The number of living stems on the date of measurement.
\item \code{status}: Indicates the status of the entire tree. For example, if any stem
is alive, the tree is alive; if every stem is dead, the tree is dead:
\itemize{
\item \code{A}: Aalive.
\item \code{D}: Dead.
\item \code{M}: Missing.
\item \code{P}: Prior. It indicates a tree had not yet recruited at this census.
\item \code{G}: Gone. xxx to complete.
}
\item \code{date}: The julian date, for date arithmetic.
\item \code{agb}: Above-ground-biomass of all stems on the tree, in Mg (equal to
metric tons or 106 g). Note that \code{agb = 0} for dead trees. FIXME: This
variable should probably be removed. It doesn't exist in the output of
rtbl().
}
}

\section{Definition of Variables in stem tables}{

Compared to tree tables, the the variables are and mean the same. But notice
some differences in \code{DFstatus} and \code{agb}; also notice the additional
variables \code{codes} and \code{countPOM} and that the variable \code{nostems} does not
exist for stem tables.
\itemize{
\item \code{treeID}: See tree table.
\item \code{stemID}: See tree table.
\item \code{tag}: See tree table.
\item \code{StemTag}: See tree table.
\item \code{sp}: See tree table.
\item \code{quadrat}: See tree table.
\item \code{gx}: See tree table.
\item \code{gy}: See tree table.
\item \code{MeasureID}: The unique identifier of a single measurement in the CTFS
database. FIXME: Missing in \code{stem6_1ha_bci}
\item \code{CensusID}: See tree table. FIXME: Misisng from \code{stem6_1ha_bci}.`
\item \code{dbh}: Diameter of the stem.
\item \code{pom}: See tree table. FIXME: Misisng from \code{stem6_1ha_bci}.
\item \code{hom}: See tree table.
\item \code{ExactDate}: See tree table.
\item \code{DFstatus}: The codes are the same as in tree tables except that stems may
have the additional status \code{G} (gone). \code{G} applies when a tree is alive and a
stem which formerly had a measurement does not in this census. This happens
in some plots, where individual stems do not have tags and cannot be tracked.
\item \code{codes}: The codes for the measurement as recorded in the field.
\item \code{countPOM}: The number of POMs (HOMs) for the same stem in this census.
FIXME: Misisng from \code{stem6_1ha_bci}.
\item \code{status}: See tree table.
\item \code{date}: See tree table.
\item \code{agb}: Same, but notice that some may be NA. FIXME: Remove from
\code{stem6_1ha_bci}. It shouldn't be here as it is not in the output of \code{rtbl()}.
}
}

\seealso{
census_bci
}
