% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\docType{data}
\name{data_dictionary}
\alias{data_dictionary}
\title{Data Dictionary of the ForestGEO database.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 242 rows and 3 columns.}
\source{
\url{http://ctfs.si.edu/Public/DataDict/data_dict.php}.
}
\usage{
data_dictionary
}
\description{
This dataset is based on the Tree Database Data Dictionary available at
\url{http://ctfs.si.edu/Public/DataDict/data_dict.php}.
}
\examples{
head(data_dictionary)

# Column names vary in subtle ways (e.g. SubSpeciesID vs. SubspeciesID)
pattern <- "^subspeciesid$"
matching <- grepl(pattern, data_dictionary$column, ignore.case = TRUE)
data_dictionary[matching, ]

}
\keyword{datasets}
