% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elevation_bci}
\alias{elevation_bci}
\title{Elevation data from Barro Colorado Island.}
\format{An object of class \code{list} of length 4.}
\usage{
elevation_bci
}
\description{
This dataset can be downloaded directly from
\url{http://bit.ly/fgeo-opendata-elevation-bci}.
}
\examples{
str(elevation_bci)
}
\keyword{datasets}
