% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.opendata-package.R
\docType{package}
\name{fgeo.opendata-package}
\alias{fgeo.opendata}
\alias{fgeo.opendata-package}
\title{fgeo.opendata: Open Datasets of ForestGEO}
\description{
Open datasets of ForestGEO.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/fgeo.opendata}
  \item Report bugs at \url{https://github.com/forestgeo/fgeo.opendata/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

Other contributors:
\itemize{
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder, funder]
}

}
\keyword{internal}
