% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-luquillo-other.R
\docType{data}
\name{luquillo_elevation}
\alias{luquillo_elevation}
\title{Elevation data from Luquillo.}
\format{An object of class \code{list} of length 4.}
\usage{
luquillo_elevation
}
\description{
This dataset can be downloaded directly from
\url{http://bit.ly/fgeo-data-luquillo-elevation}.
}
\examples{
str(luquillo_elevation, , give.attr = FALSE)

sapply(luquillo_elevation, class)

str(luquillo_elevation$col, give.attr = FALSE)
}
\seealso{
\url{https://forestgeo.si.edu/sites/north-america/luquillo}.

Other additional datasets: \code{\link{luquillo_habitat}},
  \code{\link{luquillo_species}}
}
\concept{additional datasets}
\keyword{datasets}
