% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-luquillo-other.R
\docType{data}
\name{luquillo_habitat}
\alias{luquillo_habitat}
\title{Habitat data from luquillo.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6565 rows and 3 columns.}
\usage{
luquillo_habitat
}
\description{
Habitat data from luquillo created by dividing elevation data in 5 elevation
chunks.
}
\examples{
str(luquillo_habitat)
}
\keyword{datasets}
