% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-luquillo-view.R
\docType{data}
\name{luquillo_taxa}
\alias{luquillo_taxa}
\title{Complete ViewTaxonomy (taxa) table from Luquillo.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 163 rows and 21 columns.}
\usage{
luquillo_taxa
}
\description{
Compared to the raw data in text format, this dataset replaced NULL values
with missing values, and its columns have a specific type.
}
\section{Acknowledgments}{

Thanks to Jess Zimmerman for sharing data from Luquillo.
}

\examples{
str(luquillo_taxa, give.attr = FALSE)
}
\seealso{
\itemize{
\item \link{data_dictionary}
\item \url{http://bit.ly/fgeo-data-luquillo-taxa"}.
\item \url{https://forestgeo.si.edu/sites/north-america/luquillo}.
}
}
\keyword{datasets}
