% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-luquillo-view.R
\docType{data}
\name{luquillo_vft}
\alias{luquillo_vft}
\alias{luquillo_vft_1ha}
\alias{luquillo_vft_random}
\title{ViewFullTable (vft) from Luquillo.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 38803 rows and 32 columns.}
\usage{
luquillo_vft_1ha

luquillo_vft_random
}
\description{
Slightly cleaned versions from a ViewFullTable table, from census 2005 and
census 2010 (see columns definition at \link{vft_description}):
\itemize{
\item \code{luquillo_vft_1ha}: All threes in one hectare (coordinates: \code{PX} between
(>=) 100 and 200; and \code{PY} between 400 and 500.
\item \code{luquillo_vft_random}: One thousand randomly-chosen trees (from the entire
plot).
}
}
\section{Acknowledgements}{

Thanks to Jess Zimmerman for sharing data from Luquillo.
}

\examples{
str(luquillo_vft_1ha, give.attr = FALSE)

str(luquillo_vft_random, give.attr = FALSE)

show_some_values_of_each_column <- function(x) head(unique(x))
lapply(luquillo_vft_random, show_some_values_of_each_column)
}
\seealso{
\link{vft_description}, \link{data_dictionary},
\url{http://bit.ly/fgeo-data-luquillo-vft-random},
\url{http://bit.ly/fgeo-data-luquillo-vft-1ha}.
}
\keyword{datasets}
