% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-luquillo-view.R
\docType{data}
\name{luquillo_vft_4quad}
\alias{luquillo_vft_4quad}
\title{Subset of a ViewFullTable (vft) from Luquillo.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3302 rows and 32 columns.}
\usage{
luquillo_vft_4quad
}
\description{
Subset of a ViewFullTable (vft) from Luquillo. Included are all trees in four
quadrats (coordinates: \code{PX} between (>=) 100 and 140; and \code{PY} between 400
and 440) of cencuses four through 6. For bigger datasets, including censuses
one through six, see links under section See also.
}
\details{
Compared to the raw data in text format, this dataset replaced NULL values
with missing values, and its columns have a specific type.
}
\section{Acknowledgements}{

Thanks to Jess Zimmerman for sharing data from Luquillo.
}

\examples{
str(luquillo_vft_4quad, give.attr = FALSE)

show_some_values_of_each_column <- function(x) head(unique(x))
lapply(luquillo_vft_4quad, show_some_values_of_each_column)
}
\seealso{
\link{vft_description}, \link{data_dictionary},
\url{http://bit.ly/fgeo-data-luquillo-vft-random},
\url{http://bit.ly/fgeo-data-luquillo-vft-1ha}.
}
\keyword{datasets}
