% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vft_1ha_bci}
\alias{vft_1ha_bci}
\title{All threes in one hectare from Barro Colorado Island..}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10543 rows and 32 columns.}
\usage{
vft_1ha_bci
}
\description{
One hectare of data for hectare: \code{px >= 700} and \code{px < 800}; and \code{py >=100}
and \code{py < 200}. The data includes only those records from census 2005 and
census 2010.
}
\details{
This is a slightly cleaned version from a ViewFullTable table.
For a raw .csv version see \url{http://bit.ly/fgeo-opendata-vft-1ha-bci}.
}
\examples{
str(vft_1ha_bci)
}
\keyword{datasets}
