% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-view.R
\docType{data}
\name{vft_bci}
\alias{vft_bci}
\alias{vft_1ha_bci}
\alias{vft_random_bci}
\title{ViewFullTable (vft) from Barro Colorado Island (bci).}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10543 rows and 32 columns.}
\usage{
vft_1ha_bci

vft_random_bci
}
\description{
Slightly cleaned versions from a ViewFullTable table, from census 2005 and
census 2010 (see columns definition at \link{vft_description}):
\itemize{
\item \code{vft_1ha_bci}: All threes in one hectare (coordinates: \code{px >= 700}, \code{px < 800}, \code{py >=100}, \code{py < 200}). Raw .csv file available at
\url{http://bit.ly/fgeo-opendata-vft-1ha-bci}.
\item \code{vft_random_bci}: One thousand randomly-chosen trees (from the whole plot).
Raw .csv file available at \url{http://bit.ly/fgeo-opendata-vft-random-bci}).
}
}
\examples{
str(vft_1ha_bci, give.attr = FALSE)

str(vft_random_bci, give.attr = FALSE)

show_some_values_of_each_column <- function(x) head(unique(x))
lapply(vft_random_bci, show_some_values_of_each_column)
}
\seealso{
\link{vft_description}, \link{unique_id}.
}
\keyword{datasets}
