% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-view.R
\docType{data}
\name{vft_description}
\alias{vft_description}
\title{General description of a ViewFullTable (vft).}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32 rows and 2 columns.}
\usage{
vft_description
}
\description{
Reproduced from \url{http://bit.ly/vft-definitions}:

\emph{ViewFullTable} table. Merges all the relevant variables from the
corresponding tables in the database back into a flat file. Each record
represents one measurement and/or attribute of one stem of one tree of one
plot from one census. Because buttresses can grow, for trees with buttresses,
a stem may be measured at more than one height. So some stems may have more
than one measurement in a census, but at different hom.
}
\examples{
vft_description
}
\keyword{datasets}
