% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vft_random_bci}
\alias{vft_random_bci}
\title{One Thousand trees selected at random from Barro Colorado Island.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2257 rows and 32 columns.}
\usage{
vft_random_bci
}
\description{
One Thousand, randomly chosen, trees from Barro Colorado Island (from the
whole plot). Only included data from census 2005 and census 2010.
}
\details{
This is a slightly cleaned version from a ViewFullTable table.
For a raw .csv version see \url{http://bit.ly/fgeo-opendata-vft-random-bci}.
}
\examples{
str(vft_random_bci)
}
\keyword{datasets}
