% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-census.R
\docType{data}
\name{census_bci}
\alias{census_bci}
\alias{bci_stem6_1ha}
\alias{bci_stem7_1ha}
\alias{bci_tree6_1ha}
\alias{bci_tree7_1ha}
\alias{bci_stem6_random}
\alias{bci_stem7_random}
\alias{bci_tree6_random}
\alias{bci_tree7_random}
\title{Tree and stem tables from Barro Colorado Island (bci).}
\format{See examples.}
\usage{
bci_stem6_1ha

bci_stem7_1ha

bci_tree6_1ha

bci_tree7_1ha

bci_stem6_random

bci_stem7_random

bci_tree6_random

bci_tree7_random
}
\description{
For a general description see \link{census_description}. These datasets are tree
and stem tables of the same trees selected to create the tables \link{bci_vft_1ha}
and \link{bci_vft_random}. In the name of the tree and stem tables documented
here, the number after \code{stem} or \code{census} matches a value of the column
\code{PlotCensusNumber} in the corresponding ViewFullTable.
}
\examples{
str(bci_stem6_1ha, give.attr = FALSE)

str(bci_tree7_1ha, give.attr = FALSE)

str(bci_stem7_random, give.attr = FALSE)

str(bci_tree6_random, give.attr = FALSE)

}
\seealso{
\link{bci_vft_1ha}, \link{bci_vft_random}, \link{census_description}.
}
\keyword{datasets}
