% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-census.R
\name{census_description}
\alias{census_description}
\title{General description of tree and stem tables.}
\description{
Each tree and stem table is a dataframes that stores data of one census of
one plot. That is, for any given plot one tree table stores data recorded
during one specific census, and each row records the data from one specific
tree. And for any given plot one stem table stores data recorded during one
specific census , and each row records the data from one specific stem. For
example, if plot X was censused twice, then it has two tree tables and two
stem tables.
}
\details{
Each tree or stem table has two remarkable properties:
\itemize{
\item The number of rows equals the number of trees or stems that were ever found
in the plot (in any census). This implies that the number of rows is
identical to that of any other tree or stem table of the same plot; also,
that trees or stems appear in tree or stem tables before they recruit, while
they are alive, and after they die.
\item The order of the rows is identical to that of any other tree or stem
table of the same plot. This implies that you can compare data across
censuses even if the row order is important for your analyses. Also, you can
easily repeat analyses with different censuses with minimal effort. For
example, this structure makes helps to calculate demographic rates between
any pair of censuses.
}
}
\section{Definition of Variables in tree tables}{

\itemize{
\item \code{StemTag}: FIXME: RELOCATE. Tag number on the individual stem, if present.
\item \code{treeID}: The unique tree identifier in CTFS database. Useful to be certain
in matching trees.
\item \code{tag}: Tag number used in the field.
\item \code{sp}: The species mnemonic. To get full species names, the taxonomy table
must be downloaded from the CTFS database.
\item \code{quadrat}: Quadrat designation.
\item \code{gx}: The x coordinate within the plot, relative to one edge of the plot.
\item \code{gy}: The y plot coordinate.
\item \code{stemID}: The unique stem identifier in CTFS database. Useful to be certain
in matching stems.
\item \code{dbh}: Diameter of one stem on the tree, the stem whose stemID is given.
\item \code{hom}: The height-of-measure, identical to pom but a numeric variable with
full precision.
\item \code{ExactDate}: The date on which a tree was measured.
\item \code{codes}: The codes for the measurement as recorded in the field.
\item \code{date}: The julian date, for date arithmetic.
\item \code{status}: Indicates the status of the entire tree. For example, if any stem
is alive, the tree is alive; if every stem is dead, the tree is dead:
\itemize{
\item \code{A}: Alive.
\item \code{D}: Dead.
\item \code{M}: Missing.
\item \code{P}: Prior. It indicates a tree had not yet recruited at this census.
\item \code{G}: Gone. xxx to complete.
}
\item \code{nostems}: The number of living stems on the date of measurement.
}
}

\section{Definition of Variables in stem tables}{

Compared to tree tables, the the columns of stem tables and mean the same,
except when specified below:
\itemize{
\item \code{treeID}, \code{stemID}, \code{tag}, \code{StemTag}: See tree table.
\item \code{sp}, \code{quadrat}, \code{gx}, \code{gy}. : See tree table.
\item \code{dbh}: Diameter of the stem.
\item \code{hom}, \code{ExactDate}: See tree table.
\item \code{DFstatus}: FIXME: Check with Suzanne if it should be removed.
\item \code{codes}: The codes for the measurement as recorded in the field.
\item \code{date} \code{status}: See tree table.
}
}

\seealso{
\link{census_bci}, \link{data_dictionary}.
}
