% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-luquillo-other.R
\docType{data}
\name{luquillo_habitat}
\alias{luquillo_habitat}
\title{Habitat data from luquillo.}
\format{An object of class \code{fgeo_habitat} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 503 rows and 3 columns.}
\usage{
luquillo_habitat
}
\description{
Habitat data from luquillo created by dividing elevation data in multiple
chunks.
}
\examples{
str(luquillo_habitat)
}
\seealso{
\itemize{
\item luquillo_elevation.
\item \code{fgeo.tool::fgeo_habitat()}.
\item \url{https://forestgeo.si.edu/sites/north-america/luquillo}.
}
}
\keyword{datasets}
