% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-luquillo-other.R
\docType{data}
\name{luquillo_species}
\alias{luquillo_species}
\title{SPecies table from Luquillo.}
\format{An object of class \code{data.frame} with 84 rows and 10 columns.}
\usage{
luquillo_species
}
\description{
SPecies table from Luquillo.
}
\keyword{datasets}
