% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-luquillo-view.R
\docType{data}
\name{luquillo_taxa}
\alias{luquillo_taxa}
\title{Complete ViewTaxonomy (taxa) table from Luquillo.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 163 rows and 21 columns.}
\usage{
luquillo_taxa
}
\description{
Complete ViewTaxonomy (taxa) table from Luquillo.
}
\section{Acknowledgements}{

Thanks to Jess Zimmerman for sharing data from Luquillo.
}

\examples{
str(luquillo_taxa, give.attr = FALSE)
}
\keyword{datasets}
