% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxa_bci}
\alias{taxa_bci}
\title{Complete ViewTaxonomy table from Barro Colorado Island.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1428 rows and 21 columns.}
\usage{
taxa_bci
}
\description{
Slightly cleaned version of a ViewTaxonomy table. For a raw .csv version see
\url{http://bit.ly/fgeo-opendata-taxa-bci}.
}
\keyword{datasets}
