% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-other.R
\docType{data}
\name{unique_id}
\alias{unique_id}
\title{Unique identifiers of the most commonly used tables.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 2 columns.}
\usage{
unique_id
}
\description{
Unique identifiers of the most commonly used tables.
}
\examples{
unique_id
}
\keyword{datasets}
